problems/0150-evaluate-reverse-polish-notation/eval_RPN.go
package _150_evaluate_reverse_polish_notation

import "strconv"

func evalRPN(tokens []string) int {
	operations := map[string]func(int, int) int{
		"+": func(a, b int) int { return a + b },
		"-": func(a, b int) int { return a - b },
		"*": func(a, b int) int { return a * b },
		"/": func(a, b int) int { return a / b },
	}

	var stack []int
	for _, token := range tokens {
		if op, ok := operations[token]; ok {
			a, b := stack[len(stack)-2], stack[len(stack)-1]
			stack = stack[:len(stack)-2]
			stack = append(stack, op(a, b))
		} else {
			n, _ := strconv.Atoi(token)
			stack = append(stack, n)
		}
	}
	return stack[0]
}
